dae/requirements.txt
torch==2.1.2
scikit-learn==1.4.0
scipy==1.11.4
numpy==1.26.3
pandas==2.2.0
hydra-core==1.3.2
tqdm